move createTransactionAndFirstSubmission test to describe block in name test
