+++ src/presentation/test/integration/registration/name.test.ts
@@ -1,4 +1,5 @@
 import request from "supertest";
+import { Request, Response } from "express";
 import {
   NameEndingType,
   PartnershipType
@@ -8,16 +9,22 @@ import enTranslationText from "../../../../../locales/en/translations.json";
 import cyTranslationText from "../../../../../locales/cy/translations.json";
 import app from "../app";
 import {
+  EMAIL_URL,
   NAME_URL,
   NAME_WITH_IDS_URL
 } from "../../../controller/registration/url";
 import { appDevDependencies } from "../../../../config/dev-dependencies";
-import { APPLICATION_CACHE_KEY_PREFIX_REGISTRATION } from "../../../../config/constants";
+import {
+  APPLICATION_CACHE_KEY,
+  APPLICATION_CACHE_KEY_PREFIX_REGISTRATION
+} from "../../../../config/constants";
 import RegistrationPageType from "../../../controller/registration/PageType";
 import LimitedPartnershipBuilder from "../../builder/LimitedPartnershipBuilder";
 import { getUrl, setLocalesEnabled } from "../../utils";
 
 describe("Name Page", () => {
+  const REDIRECT_URL = getUrl(EMAIL_URL);
+
   beforeEach(() => {
     setLocalesEnabled(false);
 
@@ -26,188 +33,283 @@ describe("Name Page", () => {
     appDevDependencies.cacheRepository.feedCache(null);
   });
 
-  it("should load the name page with Welsh text", async () => {
-    setLocalesEnabled(true);
+  describe("Get Name Page", () => {
+    it("should load the name page with Welsh text", async () => {
+      setLocalesEnabled(true);
 
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.LP
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.LP
+      });
+
+      const res = await request(app).get(NAME_URL + "?lang=cy");
+
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${cyTranslationText.namePage.title} - ${cyTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(cyTranslationText.namePage.title);
+      expect(res.text).toContain(cyTranslationText.namePage.whatIsName);
+      expect(res.text).toContain(cyTranslationText.namePage.nameEnding);
+      expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=cy");
+    it("should load the name page with English text", async () => {
+      setLocalesEnabled(true);
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${cyTranslationText.namePage.title} - ${cyTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(cyTranslationText.namePage.title);
-    expect(res.text).toContain(cyTranslationText.namePage.whatIsName);
-    expect(res.text).toContain(cyTranslationText.namePage.nameEnding);
-    expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
-  });
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.LP
+      });
 
-  it("should load the name page with English text", async () => {
-    setLocalesEnabled(true);
+      const res = await request(app).get(NAME_URL + "?lang=en");
 
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.LP
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${enTranslationText.namePage.title} - ${enTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(enTranslationText.namePage.title);
+      expect(res.text).toContain(enTranslationText.namePage.whatIsName);
+      expect(res.text).toContain(enTranslationText.namePage.nameEnding);
+      expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
+      expect(res.text).not.toContain("WELSH -");
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=en");
+    it("should load the name page with data from api", async () => {
+      const limitedPartnership = new LimitedPartnershipBuilder()
+        .withId(appDevDependencies.limitedPartnershipGateway.submissionId)
+        .withNameEnding(NameEndingType.LIMITED_PARTNERSHIP)
+        .build();
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${enTranslationText.namePage.title} - ${enTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(enTranslationText.namePage.title);
-    expect(res.text).toContain(enTranslationText.namePage.whatIsName);
-    expect(res.text).toContain(enTranslationText.namePage.nameEnding);
-    expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
-    expect(res.text).not.toContain("WELSH -");
-  });
+      appDevDependencies.limitedPartnershipGateway.feedLimitedPartnerships([
+        limitedPartnership
+      ]);
 
-  it("should load the name page with data from api", async () => {
-    const limitedPartnership = new LimitedPartnershipBuilder()
-      .withId(appDevDependencies.limitedPartnershipGateway.submissionId)
-      .withNameEnding(NameEndingType.LIMITED_PARTNERSHIP)
-      .build();
+      const URL = getUrl(NAME_WITH_IDS_URL);
 
-    appDevDependencies.limitedPartnershipGateway.feedLimitedPartnerships([
-      limitedPartnership
-    ]);
+      const res = await request(app).get(URL);
 
-    const URL = getUrl(NAME_WITH_IDS_URL);
-
-    const res = await request(app).get(URL);
-
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(limitedPartnership?.data?.partnership_name);
-    expect(res.text).toContain(limitedPartnership?.data?.name_ending);
-  });
-
-  it("should load the private name page with Welsh text", async () => {
-    setLocalesEnabled(true);
-
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.PFLP
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(limitedPartnership?.data?.partnership_name);
+      expect(res.text).toContain(limitedPartnership?.data?.name_ending);
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=cy");
+    it("should load the private name page with Welsh text", async () => {
+      setLocalesEnabled(true);
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${cyTranslationText.namePage.privateFund.title} - ${cyTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(cyTranslationText.namePage.privateFund.title);
-    expect(res.text).toContain(
-      cyTranslationText.namePage.privateFund.nameEnding
-    );
-    expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
-  });
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.PFLP
+      });
 
-  it("should load the private name page with English text", async () => {
-    setLocalesEnabled(true);
+      const res = await request(app).get(NAME_URL + "?lang=cy");
 
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.PFLP
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${cyTranslationText.namePage.privateFund.title} - ${cyTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(cyTranslationText.namePage.privateFund.title);
+      expect(res.text).toContain(
+        cyTranslationText.namePage.privateFund.nameEnding
+      );
+      expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=en");
+    it("should load the private name page with English text", async () => {
+      setLocalesEnabled(true);
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${enTranslationText.namePage.privateFund.title} - ${enTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(enTranslationText.namePage.privateFund.title);
-    expect(res.text).toContain(
-      enTranslationText.namePage.privateFund.nameEnding
-    );
-    expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
-    expect(res.text).not.toContain("WELSH -");
-  });
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.PFLP
+      });
 
-  it("should load the Scottish limited partnership name page with Welsh text", async () => {
-    setLocalesEnabled(true);
+      const res = await request(app).get(NAME_URL + "?lang=en");
 
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.SLP
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${enTranslationText.namePage.privateFund.title} - ${enTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(enTranslationText.namePage.privateFund.title);
+      expect(res.text).toContain(
+        enTranslationText.namePage.privateFund.nameEnding
+      );
+      expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
+      expect(res.text).not.toContain("WELSH -");
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=cy");
+    it("should load the Scottish limited partnership name page with Welsh text", async () => {
+      setLocalesEnabled(true);
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${cyTranslationText.namePage.scottish.title} - ${cyTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(cyTranslationText.namePage.scottish.title);
-    expect(res.text).toContain(cyTranslationText.namePage.scottish.nameEnding);
-    expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
-  });
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.SLP
+      });
 
-  it("should load the Scottish limited partnership name page with English text", async () => {
-    setLocalesEnabled(true);
+      const res = await request(app).get(NAME_URL + "?lang=cy");
 
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.SLP
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${cyTranslationText.namePage.scottish.title} - ${cyTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(cyTranslationText.namePage.scottish.title);
+      expect(res.text).toContain(
+        cyTranslationText.namePage.scottish.nameEnding
+      );
+      expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=en");
+    it("should load the Scottish limited partnership name page with English text", async () => {
+      setLocalesEnabled(true);
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${enTranslationText.namePage.scottish.title} - ${enTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(enTranslationText.namePage.scottish.title);
-    expect(res.text).toContain(enTranslationText.namePage.scottish.nameEnding);
-    expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
-    expect(res.text).not.toContain("WELSH -");
-  });
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.SLP
+      });
 
-  it("should load the private name Scotland page with Welsh text", async () => {
-    setLocalesEnabled(true);
+      const res = await request(app).get(NAME_URL + "?lang=en");
 
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.SPFLP
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${enTranslationText.namePage.scottish.title} - ${enTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(enTranslationText.namePage.scottish.title);
+      expect(res.text).toContain(
+        enTranslationText.namePage.scottish.nameEnding
+      );
+      expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
+      expect(res.text).not.toContain("WELSH -");
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=cy");
+    it("should load the private name Scotland page with Welsh text", async () => {
+      setLocalesEnabled(true);
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${cyTranslationText.namePage.privateFund.scottish.title} - ${cyTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(
-      cyTranslationText.namePage.privateFund.scottish.title
-    );
-    expect(res.text).toContain(cyTranslationText.namePage.whatIsNameHint);
-    expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
-  });
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.SPFLP
+      });
 
-  it("should load the private name Scotland page with English text", async () => {
-    setLocalesEnabled(true);
+      const res = await request(app).get(NAME_URL + "?lang=cy");
 
-    appDevDependencies.cacheRepository.feedCache({
-      [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
-        PartnershipType.SPFLP
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${cyTranslationText.namePage.privateFund.scottish.title} - ${cyTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(
+        cyTranslationText.namePage.privateFund.scottish.title
+      );
+      expect(res.text).toContain(cyTranslationText.namePage.whatIsNameHint);
+      expect(res.text).toContain(cyTranslationText.buttons.saveAndContinue);
     });
 
-    const res = await request(app).get(NAME_URL + "?lang=en");
+    it("should load the private name Scotland page with English text", async () => {
+      setLocalesEnabled(true);
 
-    expect(res.status).toBe(200);
-    expect(res.text).toContain(
-      `${enTranslationText.namePage.privateFund.scottish.title} - ${enTranslationText.service} - GOV.UK`
-    );
-    expect(res.text).toContain(
-      enTranslationText.namePage.privateFund.scottish.title
-    );
-    expect(res.text).toContain(enTranslationText.namePage.whatIsNameHint);
-    expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
-    expect(res.text).not.toContain("WELSH -");
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.SPFLP
+      });
+
+      const res = await request(app).get(NAME_URL + "?lang=en");
+
+      expect(res.status).toBe(200);
+      expect(res.text).toContain(
+        `${enTranslationText.namePage.privateFund.scottish.title} - ${enTranslationText.service} - GOV.UK`
+      );
+      expect(res.text).toContain(
+        enTranslationText.namePage.privateFund.scottish.title
+      );
+      expect(res.text).toContain(enTranslationText.namePage.whatIsNameHint);
+      expect(res.text).toContain(enTranslationText.buttons.saveAndContinue);
+      expect(res.text).not.toContain("WELSH -");
+    });
+  });
+
+  describe("Post Name Page", () => {
+    it("should load the name page with status 200", async () => {
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.LP
+      });
+
+      const res = await request(app).get(NAME_URL);
+
+      expect(res.status).toBe(200);
+      expect(res.text).toContain("What is the limited partnership name?");
+    });
+
+    it("should create a transaction and the first submission", async () => {
+      appDevDependencies.cacheRepository.feedCache({
+        [`${APPLICATION_CACHE_KEY_PREFIX_REGISTRATION}${RegistrationPageType.whichType}`]:
+          PartnershipType.LP
+      });
+
+      const res = await request(app).post(NAME_URL).send({
+        pageType: RegistrationPageType.name,
+        partnership_name: "Test Limited Partnership",
+        name_ending: NameEndingType.LIMITED_PARTNERSHIP,
+        partnership_type: PartnershipType.LP
+      });
+
+      expect(res.status).toBe(302);
+      expect(res.text).toContain(`Redirecting to ${REDIRECT_URL}`);
+
+      expect(appDevDependencies.cacheRepository.cache).toEqual({
+        [APPLICATION_CACHE_KEY]: {}
+      });
+    });
+
+    it("should update the submission if already exist", async () => {
+      const limitedPartnership = new LimitedPartnershipBuilder()
+        .withId(appDevDependencies.limitedPartnershipGateway.submissionId)
+        .withNameEnding(NameEndingType.LIMITED_PARTNERSHIP)
+        .build();
+
+      appDevDependencies.limitedPartnershipGateway.feedLimitedPartnerships([
+        limitedPartnership
+      ]);
+
+      const URL = getUrl(NAME_WITH_IDS_URL);
+
+      const res = await request(app).post(URL).send({
+        pageType: RegistrationPageType.name,
+        partnership_name: "Test Limited Partnership",
+        name_ending: NameEndingType.LIMITED_PARTNERSHIP,
+        partnership_type: PartnershipType.LP
+      });
+
+      expect(res.status).toBe(302);
+      expect(res.text).toContain(`Redirecting to ${REDIRECT_URL}`);
+    });
+
+    it("should return validation errors", async () => {
+      const res = await request(app).post(NAME_URL).send({
+        pageType: RegistrationPageType.name
+      });
+
+      expect(res.status).toBe(200);
+      expect(res.text).toContain("partnership_name must be less than 160");
+    });
+
+    it("should return a status 500 if page type doesn't exist - sq", async () => {
+      const res = await request(app).post(NAME_URL).send({
+        pageType: "wrong-page-type"
+      });
+
+      expect(res.status).toBe(500);
+    });
+
+    it("should call next if type in path is incorrect - sq", async () => {
+      const next = jest.fn();
+
+      await appDevDependencies.globalController.getPageRouting()(
+        {
+          path: "/limited-partnerships/wrong-type"
+        } as Request,
+        {} as Response,
+        next
+      );
+
+      expect(next).toHaveBeenCalled();
+    });
   });
 });
